Move `Mine` class to "MapObjects" folder
