+++ OPHD/Map/Tile.cpp
@@ -1,6 +1,6 @@
 #include "Tile.h"
 
-#include "../Mine.h"
+#include "../MapObjects/Mine.h"
 #include "../MapObjects/Robot.h"
 #include "../MapObjects/Structure.h"
