+++ OPHD/MapObjects/Mine.cpp
