+++ OPHD/States/MapViewStateDraw.cpp
@@ -9,9 +9,9 @@
 #include "../Constants/UiConstants.h"
 #include "../Common.h"
 #include "../Cache.h"
-#include "../Mine.h"
 #include "../StructureManager.h"
 #include "../Map/TileMap.h"
+#include "../MapObjects/Mine.h"
 
 #include <NAS2D/Utility.h>
 #include <NAS2D/Renderer/Renderer.h>
